+++ cmd/simclient/constants.go
@@ -4,7 +4,7 @@ import "time"
 
 const (
 	defaultAPIBase       = "http://localhost:8080"
-	defaultViewerBase    = "http://localhost:8081"
+	defaultViewerBase    = "http://localhost:8080"
 	defaultChunks        = 12
 	defaultInterval      = 5 * time.Second
 	defaultMediaType     = "audio"